Add a boundary check for safety and use snprintf() to construct the IP address strings instead of strl{cpy,cat}().  Also make the function return a valid string even in failure case.
input deraadt
ok deraadt

+++ usr.sbin/radiusd/util.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD$	*/
+/*	$OpenBSD: util.h,v 1.1 2015/07/21 04:06:04 yasuoka Exp $	*/
 
 #ifndef RADIUSD_UTIL_H
 #define	RADIUSD_UTIL_H
@@ -25,7 +25,7 @@
 __BEGIN_DECLS
 
 int		 addrport_parse(const char *, int, struct addrinfo **);
-const char	*addrport_tostring (struct sockaddr *, socklen_t, char *, int);
+const char	*addrport_tostring (struct sockaddr *, socklen_t, char *, size_t);
 
 __END_DECLS
 #endif
